+++ src/mainReadout.cxx
@@ -16,6 +16,7 @@
 
 #include <Common/Configuration.h>
 #include <Common/Fifo.h>
+#include <Common/LineBuffer.h>
 #include <Common/MemPool.h>
 #include <Common/Thread.h>
 #include <Common/Timer.h>
@@ -92,6 +93,9 @@
 #ifdef WITH_OCC
 #include <OccInstance.h>
 #include <RuntimeControlledObject.h>
+#else
+#define OCC_CONTROL_PORT_ENV ""
+#define OCC_ROLE_ENV ""
 #endif
 
 // namespace used
@@ -171,6 +175,12 @@ class Readout
   int cfgVerbose = 0;
   int cfgMaxMsgError; // maximum number of error messages before stopping run
   int cfgMaxMsgWarning; // maximum number of warning messages before stopping run
+  int cfgCustomCommandsEnabled = 0; // when set, a sub-process bash is launched to execute custom commands
+  std::map<std::string,std::string> customCommands; // map of state / command pairs to be executed
+  pid_t customCommandsShellPid = 0; // pid of shell for custom commands
+  int customCommandsShellFdIn = -1; // input to shell
+  int customCommandsShellFdOut = -1; // output from shell
+  void executeCustomCommand(const char *stateChange); // execute a custom command for given state transition, if any
 
   // runtime entities
   std::vector<std::unique_ptr<Consumer>> dataConsumers;
@@ -280,6 +290,7 @@ int Readout::init(int argc, char* argv[])
     cfgDefaults.getOptionalValue<std::string>(cfgDefaultsEntryPoint + ".statsPublishAddress", cfgStatsPublishAddress, cfgStatsPublishAddress);
     cfgDefaults.getOptionalValue<double>(cfgDefaultsEntryPoint + ".statsPublishInterval", cfgStatsPublishInterval, cfgStatsPublishInterval);
     cfgDefaults.getOptionalValue<std::string>(cfgDefaultsEntryPoint + ".db", cfgDatabaseCxParams);
+    cfgDefaults.getOptionalValue<int>(cfgDefaultsEntryPoint + ".customCommandsEnabled", cfgCustomCommandsEnabled);
   }
   catch(...) {
     //initLogs.push_back({LogWarningSupport_(3100), std::string("Error loading defaults")});
@@ -408,12 +419,47 @@ int Readout::init(int argc, char* argv[])
       try {
         dbHandle=std::make_unique<ReadoutDatabase>(cfgDatabaseCxParams.c_str(), cfgVerbose, dbLog);
 	if (dbHandle == nullptr) { throw __LINE__; }
-	theLog.log(LogInfoDevel_(3012), "Database connected ");
+	theLog.log(LogInfoDevel_(3012), "Database connected");
       }
       catch(...) {
         theLog.log(LogWarningDevel_(3242), "Failed to connect database");
       }
     #endif
+  }
+
+  // init shell for custom commands
+  if (cfgCustomCommandsEnabled) {
+    for (;;) {
+      int p_stdin[2], p_stdout[2];
+      pid_t pid;
+
+      if (pipe(p_stdin) != 0 || pipe(p_stdout) != 0) break;
+      pid = fork();
+
+      if (pid < 0) {
+        break;
+      } else if (pid == 0) {
+        dup2(p_stdin[0], STDIN_FILENO);
+        dup2(p_stdout[1], STDOUT_FILENO);
+	close(p_stdin[0]);
+	close(p_stdin[1]);
+	close(p_stdout[0]);
+	close(p_stdout[1]);
+        execl("/bin/bash", "bash", NULL);
+        exit(1);
+      }
+      close(p_stdin[0]);
+      close(p_stdout[1]);
+      customCommandsShellFdIn = p_stdin[1];
+      customCommandsShellFdOut = p_stdout[0];
+      customCommandsShellPid = pid;
+      break;
+    }
+    if (customCommandsShellPid) {
+      theLog.log(LogInfoDevel_(3013), "Shell started for custom commands - pid %d", (int)customCommandsShellPid);
+    } else {
+      cfgCustomCommandsEnabled = 0;
+    }
   }
 
   return 0;
@@ -535,6 +581,22 @@ int Readout::configure(const boost::property_tree::ptree& properties)
   }
 
   // extract optional configuration parameters
+  // configuration parameter: | readout | customCommands | string | | List of key=value pairs defining some custom shell commands to be executed at before/after state change commands. |
+  if (customCommandsShellPid) {
+    std::string cfgCustomCommandsList;
+    customCommands.clear();
+    cfg.getOptionalValue<std::string>("readout.customCommands", cfgCustomCommandsList);
+    if (getKeyValuePairsFromString(cfgCustomCommandsList, customCommands)) {
+      theLog.log(LogWarningDevel_(3102), "Failed to parse custom commands");
+      customCommands.clear();
+    } else {
+      theLog.log(LogInfoDevel_(3013), "Registered custom commands:");
+      for (const auto &kv : customCommands) {
+	theLog.log(LogInfoDevel_(3013), "%s : %s", kv.first.c_str(), kv.second.c_str());
+      }
+    }
+  }
+
   // configuration parameter: | readout | exitTimeout | double | -1 | Time in seconds after which the program exits automatically. -1 for unlimited. |
   cfgExitTimeout = -1;
   cfg.getOptionalValue<double>("readout.exitTimeout", cfgExitTimeout);
@@ -985,6 +1047,9 @@ int Readout::start()
   logbookTimer.reset(cfgLogbookUpdateInterval * 1000000);
   maxTimeframeId = 0;
 
+  // execute custom command
+  executeCustomCommand("preSTART");
+
   // cleanup exit conditions
   ShutdownRequest = 0;
 
@@ -1035,10 +1100,14 @@ int Readout::start()
   std::function<void(void)> l = std::bind(&Readout::loopRunning, this);
   runningThread = std::make_unique<std::thread>(l);
 
+  // execute custom command
+  executeCustomCommand("postSTART");
+
   theLog.log(LogInfoSupport_(3005), "Readout completed START");
   gReadoutStats.counters.state = stringToUint64("running");
   gReadoutStats.counters.notify++;
   gReadoutStats.publishNow();
+
   return 0;
 }
 
@@ -1172,6 +1241,9 @@ int Readout::stop()
   gReadoutStats.counters.notify++;
   gReadoutStats.publishNow();
 
+  // execute custom command
+  executeCustomCommand("preSTOP");
+
   // raise flag
   stopTimer.reset(cfgFlushEquipmentTimeout * 1000000); // add a delay before stopping aggregator - continune to empty FIFOs
   isRunning = 0;
@@ -1231,6 +1303,9 @@ int Readout::stop()
     NumberOfBytesToString(gReadoutStats.counters.bytesReadout.load(), "bytes").c_str(),
     NumberOfBytesToString(gReadoutStats.counters.bytesRecorded.load(),"bytes").c_str()
   );
+
+  // execute custom command
+  executeCustomCommand("postSTOP");
 
   theLog.log(LogInfoSupport_(3005), "Readout completed STOP");
   gReadoutStats.counters.state = stringToUint64("ready");
@@ -1331,6 +1406,43 @@ Readout::~Readout() {
     d->abortThread();
   }
   readoutDevices.clear(); // after aggregator, because they own the data blocks
+
+  if (customCommandsShellPid) {
+    if (cfgVerbose) {
+      theLog.log(LogInfoDevel_(3013), "Closing custom command shell");
+    }
+    if (customCommandsShellFdIn >= 0) {
+      close(customCommandsShellFdIn);
+    }
+    if (customCommandsShellFdOut >= 0) {
+      close(customCommandsShellFdOut);
+    }
+    kill(customCommandsShellPid, SIGKILL);
+  }
+
+  #ifdef WITH_DB
+  dbHandle = nullptr;
+  #endif
+}
+
+void Readout::executeCustomCommand(const char *stateChange) {
+  if (customCommandsShellPid) {
+    auto it = customCommands.find(stateChange);
+    if (it != customCommands.end()) {
+      theLog.log(LogInfoDevel_(3013), "Executing custom command for %s : %s", it->first.c_str(), it->second.c_str());
+      std::string cmd = it->second + "\n";
+      write(customCommandsShellFdIn, cmd.c_str(), cmd.length());
+      LineBuffer b;
+      const int cmdTimeout = 5000; // 5s timeout
+      b.appendFromFileDescriptor(customCommandsShellFdOut, cmdTimeout);
+      std::string result;
+      if (b.getNextLine(result) == 0) {
+        theLog.log(LogInfoDevel_(3013), "Command executed: %s", result.c_str());
+      } else {
+        theLog.log(LogInfoDevel_(3013), "Unknown command result");
+      }
+    }
+  }
 }
 
 #ifdef WITH_OCC
@@ -1771,6 +1883,10 @@ int main(int argc, char* argv[])
   gReadoutStats.stopPublish();
 
   theReadout = nullptr;
+
+  #ifdef WITH_DB
+  mysql_library_end();
+  #endif
 
   theLog.log(LogInfoSupport_(3001), "Readout process exiting");
   return 0;